types.ts 생성
