+++ src/components/Login.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
 import styled from "styled-components";
-export interface MessageRec{
-  message:string;
-  timestamp:any;
-  user:string;
-  userImage:string;
-}
+import {MessageRec} from './types';
 function Message({message, timestamp, user, userImage}:MessageRec) {
   return (
     <MessageContainer>
